+++ tests/test_methods/test_local.py
@@ -2,5 +2,4 @@ import pytest
 from votesmart.methods.local import *
 
 def test_Local():
-    with pytest.raises(NotImplementedError):
-        method = Local(api_instance='test')
+    method = Local(api_instance='test')