+++ tests/test_methods/test_measure.py
@@ -2,5 +2,4 @@ import pytest
 from votesmart.methods.measure import *
 
 def test_Measure():
-    with pytest.raises(NotImplementedError):
-        method = Measure(api_instance='test')
+    method = Measure(api_instance='test')